requirements.txt
cudatoolkit==11.3.1
h5py==3.6.0
matplotlib==3.5.1
numpy==1.22.3
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
shap==0.40.0
pytables==3.7.0
pytorch==1.11.0
tqdm==4.62.3
xgboost==1.6.1
hyperopt==0.2.7